dynamic excel generator function included in class and model fields from frntend
