+++ .isort.cfg
@@ -1,2 +1,2 @@
 [settings]
-known_third_party = asgiref,celery,channels,ckeditor,corsheaders,dj_rest_auth,django,django_celery_beat,django_filters,drf_yasg,environ,fcm_django,firebase_admin,jwt,nested_admin,pyotp,pytz,requests,rest_framework,segno
+known_third_party = asgiref,celery,channels,ckeditor,corsheaders,dj_rest_auth,django,django_celery_beat,django_filters,drf_yasg,environ,fcm_django,firebase_admin,jwt,nested_admin,pyotp,pytz,requests,rest_framework,segno,xlsxwriter
