+++ common/api/views.py
@@ -1,4 +1,4 @@
-from rest_framework.generics import CreateAPIView, ListAPIView, UpdateAPIView
+from rest_framework.generics import CreateAPIView, GenericAPIView, UpdateAPIView
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
@@ -24,12 +24,13 @@ class BaseCreatorUpdateAPIView(UpdateAPIView):
         )
 
 
-class BaseReportGeneratorAPIView(ListAPIView):
+class BaseReportGeneratorAPIView(GenericAPIView):
     filter_backends = None
     search_fields = None
     queryset = None
     filterset_class = None
     model_name = None
+    serializer_class = None
 
     def get_context_data(self, **kwargs):
         ctx = super().get_context_data(**kwargs)
@@ -37,11 +38,14 @@ class BaseReportGeneratorAPIView(ListAPIView):
             ctx["model_name"] = self.model_name
         return ctx
 
-    def get(self, request):
+    def post(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
         filtered_data = self.filter_queryset(self.get_queryset())
         excelcelery.delay(
-            list(filtered_data.values_list("pk", flat=True)),
+            list(serializer.data["model_fields"]),
             self.model_name,
+            list(filtered_data.values_list("pk", flat=True)),
             request.user.id,
         )
         return Response({"msg": "Your will be notified after your file is ready."})
