+++ common/report.py
@@ -1,12 +1,25 @@
+import os
+from pathlib import Path
+
+import xlsxwriter
+from django.conf import settings
+from django.contrib.auth import get_user_model
+
+from common.utils import get_random_string
+from report.models import GeneratedReport
+
+User = get_user_model()
+
 
 class BaseDynamicTableData:
     field_to_header_names = None
+    model = None
 
-    def __init__(self, model_fields, queryset, worksheet):
+    def __init__(self, queryset, user_id, model_fields):
         self.model_fields = model_fields
-        self.worksheet = worksheet
+        self.user_id = user_id
         self.queryset = queryset
-    
+
     def get_header_names_from_field_names(self):
         header_list = []
         for key, value in self.field_to_header_names.items():
@@ -16,25 +29,43 @@ class BaseDynamicTableData:
         return header_list
 
     def generate_report(self):
-        # retrieve ExamThroughEnrollment model data
-        queryset = self.queryset
+        # retrieve model data
+        queryset = self.model.objects.filter(id__in=self.queryset)
 
-        worksheet = self.worksheet
-        worksheet.write(0, 0,'S.No')
+        user = User.objects.get(id=self.user_id)
+        media_path = f"reports/{user.username}"
+        base_path = os.path.join(settings.BASE_DIR, f"media/{media_path}")
+        os.makedirs(base_path, exist_ok=True)
+
+        filename = get_random_string()
+        p = Path(f"{base_path}/{filename}.xlsx")
+        workbook = xlsxwriter.Workbook(p)
+        worksheet = workbook.add_worksheet("report")
+
+        worksheet.write(0, 0, "S.No")
         i = 1
         # For header names
         for head_name in self.get_header_names_from_field_names():
             worksheet.write(0, i, head_name)
-            i= i+1
+            i = i + 1
 
         # For data
         col = 0
         row = 1
         for linea in queryset:
-            worksheet.write(row, col, row)  #For serial number.
+            worksheet.write(row, col, row)  # For serial number.
             for field_name in self.model_fields:
-                col=col+1
-                worksheet.write(row, col, self.get_values_from_fields(field_name, linea))
-            row=row+1
+                col = col + 1
+                worksheet.write(
+                    row, col, self.get_values_from_fields(field_name, linea)
+                )
+            row = row + 1
             col = 0
-        return worksheet
+
+        workbook.close()
+
+        GeneratedReport.objects.create(
+            created_by=user,
+            updated_by=user,
+            report_file=f"{media_path}/{filename}.xlsx",
+        )
