+++ enrollments/api/utils.py
@@ -1,8 +1,6 @@
-from django.contrib.auth import get_user_model
 from rest_framework import serializers
 
 from enrollments.models import (
-    CourseThroughEnrollment,
     EnrollmentStatus,
     ExamSession,
     ExamThroughEnrollment,
@@ -131,104 +129,3 @@ def retrieve_exam_status(self, obj):
                 return ExamStatus.IN_PROGRESS
         return ExamStatus.CREATED
     return ExamStatus.CREATED
-
-
-def dynamic_excel_generator(queryset_id, data, user_id):
-    import os
-    from pathlib import Path
-
-    import xlsxwriter
-    from django.conf import settings
-
-    from accounts.models import Profile
-    from report.models import GeneratedReport
-    from report.views import (
-        CourseThroughEnrollmentTableData,
-        ExamThroughEnrollmentTableData,
-        StudentTableData,
-    )
-
-    User = get_user_model()
-    # Create a workbook and add a worksheet.
-    user = User.objects.get(id=1)
-    media_path = f"reports/{user.username}"
-    base_path = os.path.join(settings.BASE_DIR, f"media/{media_path}")
-    os.makedirs(base_path, exist_ok=True)
-
-    filename = get_random_string()
-    p = Path(f"{base_path}/{filename}.xlsx")
-    workbook = xlsxwriter.Workbook(p)
-    worksheet = workbook.add_worksheet("report")
-    # bold = workbook.add_format({"bold": True})
-    # Some data we want to write to the worksheet.
-    # passing field names received from front-end
-
-    if data == "ExamThroughEnrollment":
-        # get model names and it correcponding headers needed in report.
-        model_fields = [
-            "enrollment",
-            "exam",
-            "selected_session",
-            "rank",
-            "score",
-            "negative_score",
-            "status",
-        ]
-        queryset = ExamThroughEnrollment.objects.filter(id__in=queryset_id)
-        report = ExamThroughEnrollmentTableData(model_fields, queryset, worksheet)
-    elif data == "CourseThroughEnrollment":
-        model_fields = [
-            "enrollment",
-            "course_name",
-            "selected_session",
-            "course_enroll_status",
-            "completed_date",
-        ]
-        queryset = CourseThroughEnrollment.objects.filter(id__in=queryset_id)
-        report = CourseThroughEnrollmentTableData(model_fields, queryset, worksheet)
-    elif data == "StudentProfile":
-        model_fields = [
-            "username",
-            "fullname",
-            "email",
-            "college_name",
-            "faculty",
-        ]
-        queryset = Profile.objects.filter(id__in=queryset_id)
-        report = StudentTableData(model_fields, queryset, worksheet)
-    worksheet = report.generate_report()
-    workbook.close()
-
-    GeneratedReport.objects.create(
-        created_by=user, updated_by=user, report_file=f"{media_path}/{filename}.xlsx"
-    )
-
-    """
-    if need to send file to front-end
-    """
-    # output.seek(0)
-    # """
-    # For testing without front-end
-    # """
-    # response = HttpResponse(
-    #     output.read(),
-    #     content_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
-    # )
-    # # response['Content-Disposition'] = "attachment; filename=report.xlsx"
-    # """
-    # To send actual xlsx file.
-    # """
-    # blob = base64.b64encode(output.read())
-    # response = HttpResponse(blob, content_type="application/ms-excel")
-    # response["Content-Disposition"] = "attachment; filename=report.xlsx"
-
-    # return response
-
-
-def get_random_string():
-    import random
-    import string
-
-    # With combination of lower and upper case
-    result_str = "".join(random.choice(string.ascii_letters) for i in range(7))
-    return result_str
