+++ exams/api/urls.py
@@ -6,7 +6,6 @@ from exams.api.views import (
     ExamPaperPreviewAPIView,
     ExamRetrieveAPIView,
     ExamRetrievePoolAPIView,
-    ExamGeneratorListAPIView
 )
 
 urlpatterns = [
@@ -27,10 +26,4 @@ urlpatterns = [
         ExamPaperPreviewAPIView.as_view(),
         name="exam-paper-preview",
     ),
-    path(
-        "generator/list/", 
-        ExamGeneratorListAPIView.as_view(),
-        name="generator-exam"
-    ),    
-
 ]
