+++ exams/api/views.py
@@ -79,17 +79,3 @@ class ExamPaperPreviewAPIView(RetrieveAPIView):
     serializer_class = ExamPaperSerializer
     permission_classes = [IsAdminUser]
     queryset = Exam.objects.all()
-
-
-class ExamGeneratorListAPIView(ListAPIView):
-    # permission_classes = [IsAuthenticated]
-    queryset = Exam.objects.all()
-    serializer_class = ExamListSerializer
-    model = Exam
-
-
-    def list(self, request, *args, **kwargs):
-        queryset = self.filter_queryset(self.get_queryset())
-        dynamic_excel_generator(self.model.__name__,queryset)
-        serializer = self.get_serializer(queryset, many=True)
-        return Response(serializer.data)
